feat: user tool settings resolution, UI fixes, and frontend build improvements
- Fix user personal API keys (tool_user_settings) being ignored during agent
  execution by threading userId through ToolInterface → Orchestrator → Tools
  and merging user settings into getEffectiveSettings() cascade
- Rename "Use global default" to "Delete settings" in ToolSettingsForm with a
  modal confirm dialog; show correct "Settings deleted." success message
- Fix vue-tsc build errors in AgentHeaderToolbar, AgentSidebar, and
  SettingsSidebar by using @ts-ignore for emit() type inference issues
- Fix no-unused-vars lint errors across 7 Vue components; fix
  no-useless-assignment in TaskChatPage.approveAll()
- Update AgentHeaderToolbar.spec.ts emit name from 'open-sidebar' to
  'openSidebar' to match new defineEmits type

+++ database/migrations/000008_create_agent_memory_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     {
         Capsule::schema()->create('agent_memory', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('agent_id');
+            $table->unsignedBigInteger('agent_id');
             $table->string('key');
             $table->text('value')->nullable();
             $table->timestamps();
